Fix Deal.II compilation problem (caused by MPI and triangle both defining REAL)
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@3434 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/src/common/ConformingTetrahedralMesh.cpp
@@ -1016,7 +1016,7 @@ void ConformingTetrahedralMesh<ELEMENT_DIM, SPACE_DIM>::ReMeshWithTriangleLibrar
 {
     struct triangulateio triangle_input;
     //struct triangulateio triangle_output;
-    triangle_input.pointlist = (REAL *) malloc(GetNumNodes() * 2 * sizeof(REAL));
+    triangle_input.pointlist = (double *) malloc(GetNumNodes() * 2 * sizeof(double));
     triangle_input.numberofpoints = GetNumNodes();
     triangle_input.numberofpointattributes = 0;
     unsigned new_index = 0;
